+++ src/types.ts
@@ -2,6 +2,7 @@ import 'ultimate-crosscode-typedefs'
 import type { Mod } from 'ultimate-crosscode-typedefs/modloader/mod'
 import type {} from 'cc-map-screenshot/src/plugin'
 import type {} from 'crossnode/crossnode.d.ts'
+import type {} from 'nax-ccuilib/src/plugin'
 
 export type Mod1 = Mod & {
     findAllAssets?(): void /* only there for ccl2, used to set isCCL3 */