+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Don't display instance labels while taking a screenshot with cc-map-screenshot
 
+### Fixed
+
+- Improve CCModManager compatibility
+
 ## [1.1.5] 2026-08-01
 
 ### Fixed